trying to comment
Im gonna get a couple hours of sleep, but here is the program, mostly commented.
As of right now, players cannot play more than one game. i have not done the memory re-allocation.
The ships are now properly destroyed when hit, and the game properly ends when all pieces are destroyed.

The two files i have been using (Player1.shp and Player2.shp) test for invalid input and previously excluded ships.
'/' in both the column and row in the file means that the ship was excluded by the user that saved the file.

I will attempt to finish commenting in the morning. If someone else can do a test plan that would be awesome